Replace internal usage of JpegLSPresetCodingParameters with jpegls_pc_parameters
The JpegLSPresetCodingParameters is a public type, used by the legacy API. Remove the internal usage of this type to make it possible to remove this type in the future.

+++ samples/convert.cpp/main.cpp
@@ -69,7 +69,7 @@ int main(const int argc, char const* const argv[])
         int near_lossless{};
         if (argc > 3)
         {
-            near_lossless = strtol(argv[3], nullptr, 10);
+            near_lossless = static_cast<int>(strtol(argv[3], nullptr, 10));
             if (near_lossless < 0 || near_lossless > 255)
             {
                 cerr << "near-lossless-value needs to be in the range [0,255]\n";
